Thread the IPG frequency into the PIT configuration
