+++ teensy4-examples/src/usb.rs
@@ -16,7 +16,6 @@ fn main() -> ! {
     // Initialize logging with the default settings
     p.log.init(Default::default());
     let mut led = p.led;
-    let mut counter = 0;
     loop {
         bsp::delay(1000);
         log::error!("Something terrible happened!");
@@ -25,14 +24,5 @@ fn main() -> ! {
         log::debug!("Sleeping for 1 second...");
         log::trace!("{} + {} = {}", 3, 2, 3 + 2);
         led.toggle().unwrap();
-        counter += 1;
-        if counter == 5 {
-            let (ccm, ccm_analog) = p.ccm.handle.raw();
-            let dcdc = p.dcdc.raw();
-            bsp::hal::set_arm_clock(600_000_000, ccm, ccm_analog, dcdc);
-            loop {
-                core::sync::atomic::spin_loop_hint();
-            }
-        }
     }
 }
